Add more balance.

+++ models/character.py
@@ -55,7 +55,7 @@ class Attribute(models.Model):
 	name = models.CharField(max_length=50)
 	owner = models.ForeignKey(Character)
 	level = models.IntegerField()
-	exp = models.IntegerField()
+	exp = models.FloatField()
 	class Meta:
 		app_label = 'kovin'
 	def __unicode__(self):
